Move computer methods to Computer class and update display_boards
