+++ lib/game.rb
@@ -1,44 +1,19 @@
 require './lib/ship'
 require './lib/board'
+require './lib/computer'
 
 class Game
   def initialize
   end
 
-  def random_coords(ship, board)
-    coords = board.cells.keys.sample(ship.length)
-
-    if board.valid_placement?(ship, coords) == false
-      coords = board.cells.keys.sample(ship.length) until
-        board.valid_placement?(ship, coords) == true
-      coords
-    end
-  end
-
-  def com_placement
-    cruiser = Ship.new(cruiser, 3)
-    submarine = Ship.new(submarine, 2)
-    com_board = Board.new(4)
-    # require 'pry'; binding.pry
-
-    com_board.place(cruiser, random_coords(cruiser, com_board))
-    com_board.place(submarine, random_coords(submarine, com_board))
-    com_board.render
-  end
-
-  def com_speaks
-    puts "I have laid out my ships on the grid."
-    puts "You now need to lay out your two ships."
-    puts "The Cruiser is three units long and the Submarine is two units long."
-  end
-
-  def display_boards
+  def display_boards(com)
     system "clear"
     system "cls"
     puts '=============COMPUTER BOARD============='
-    puts com_placement
+    # require 'pry'; binding.pry
+    puts com.com_board.render
     puts '==============PLAYER BOARD=============='
-    puts com_placement
+    # puts player.player_board.render
   end
 
   def play
@@ -91,11 +66,15 @@ class Game
     end
 
     if play_mode == '4'
+      com = Computer.new(4)
+      puts display_boards(com)
       # We start the turing game from here
-      display_boards
+
     end
 
     if play_mode == 't'
+      com = Computer.new(10)
+      puts display_boards(com)
       # We start the traditional game from here
     end
