+++ spec/game_spec.rb
@@ -19,8 +19,8 @@ RSpec.describe Game do
 
     it 'returns the board when the com places ships' do
       game = Game.new
-      require 'pry'; binding.pry
-      expect(game.com_placement).to be_a(Board)
+      # require 'pry'; binding.pry
+      expect(game.com_board).to be_a(Board)
     end
   end
 end
